requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
defusedxml==0.7.1
Django==6.0.4
django-bootstrap5==26.2
fonttools==4.62.1
fpdf2==2.8.7
idna==3.13
pillow==12.2.0
requests==2.33.1
sqlparse==0.5.5
stripe==15.1.0
typing_extensions==4.15.0
urllib3==2.6.3
